+++ crates/plotinator-logs/src/lib.rs
@@ -3,8 +3,11 @@ use std::io;
 use plotinator_log_if::prelude::*;
 
 pub mod generator;
+pub mod inclinometer_sps;
+pub mod mag_sps;
 pub mod mbed_motor_control;
 pub mod navsys;
+pub mod navsys_kitchen_sink;
 pub mod util;
 pub mod wasp200;