+++ crates/plotinator-logs/src/navsys/entries/mag.rs
@@ -15,6 +15,10 @@ pub struct MagSensor {
 }
 
 impl MagSensor {
+    pub(crate) fn timestamp(&self) -> DateTime<Utc> {
+        self.timestamp
+    }
+
     pub(crate) fn timestamp_ns(&self) -> f64 {
         self.timestamp
             .timestamp_nanos_opt()