+++ crates/plotinator-logs/src/navsys_kitchen_sink.rs
@@ -5,22 +5,23 @@ use std::{
 };
 
 use chrono::{DateTime, Utc};
-use entries::NavSysSpsEntry;
-use header::NavSysSpsHeader;
 use plotinator_log_if::{parseable::Parseable, prelude::*};
 use serde::{Deserialize, Serialize};
 
-pub(crate) mod entries;
-pub(crate) mod header;
+use crate::{
+    mag_sps::MagSps,
+    navsys::entries::{NavSysSpsEntry, tl::InclinometerEntry},
+    wasp200::Wasp200Sps,
+};
 
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
-pub struct NavSysSps {
-    header: NavSysSpsHeader,
+pub struct NavSysSpsKitchenSink {
+    first_timestamp: DateTime<Utc>,
     entries: Vec<NavSysSpsEntry>,
     raw_plots: Vec<RawPlot>,
 }
 
-impl NavSysSps {
+impl NavSysSpsKitchenSink {
     /// Read a file and attempt to deserialize a `NavSysSps` header from it
     ///
     /// Return true if a valid header was deserialized
@@ -29,7 +30,17 @@ impl NavSysSps {
             return false;
         };
         let mut reader = BufReader::new(file);
-        NavSysSpsHeader::from_reader(&mut reader).is_ok()
+
+        let valid_tilt_sensor = Self::is_reader_valid_tilt_sensor(&mut reader);
+
+        MagSps::file_is_valid(path) || Wasp200Sps::file_is_valid(path) || valid_tilt_sensor
+    }
+
+    fn is_reader_valid_tilt_sensor(reader: &mut impl io::BufRead) -> bool {
+        // If 3 entries can be read successfully then it's valid
+        InclinometerEntry::from_reader(reader).is_ok()
+            && InclinometerEntry::from_reader(reader).is_ok()
+            && InclinometerEntry::from_reader(reader).is_ok()
     }
 
     #[allow(
@@ -336,9 +347,8 @@ impl NavSysSps {
     }
 }
 
-impl fmt::Display for NavSysSps {
+impl fmt::Display for NavSysSpsKitchenSink {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        writeln!(f, "{}", self.header)?;
         for e in &self.entries {
             writeln!(f, "{e}")?;
         }
@@ -346,13 +356,13 @@ impl fmt::Display for NavSysSps {
     }
 }
 
-impl Plotable for NavSysSps {
+impl Plotable for NavSysSpsKitchenSink {
     fn raw_plots(&self) -> &[RawPlot] {
         &self.raw_plots
     }
 
     fn first_timestamp(&self) -> DateTime<Utc> {
-        self.header.first_timestamp()
+        self.first_timestamp
     }
 
     fn descriptive_name(&self) -> &str {
@@ -364,40 +374,59 @@ impl Plotable for NavSysSps {
     }
 
     fn metadata(&self) -> Option<Vec<(String, String)>> {
-        let metadata: Vec<(String, String)> = vec![
-            ("Version".into(), self.header.version().to_string()),
-            (
-                "NavSys Software rev.".into(),
-                self.header.software_revision().to_owned(),
-            ),
-            (
-                "TiltSensor ID".into(),
-                self.header.tilt_sensor_id().to_owned(),
-            ),
-        ];
-
-        Some(metadata)
+        None
     }
 }
 
-impl Parseable for NavSysSps {
-    const DESCRIPTIVE_NAME: &str = "NavSys Sps";
+impl Parseable for NavSysSpsKitchenSink {
+    const DESCRIPTIVE_NAME: &str = "NavSys Sps Kitchen Sink";
 
     fn from_reader(reader: &mut impl io::BufRead) -> io::Result<(Self, usize)> {
+        let mut entries = Vec::new();
         let mut total_bytes_read = 0;
-        let (header, bytes_read) = NavSysSpsHeader::from_reader(reader)?;
-        total_bytes_read += bytes_read;
 
-        let (entries, bytes_read) = parse_to_vec(reader);
-        total_bytes_read += bytes_read;
+        let mut first_timestamp: Option<DateTime<Utc>> = None;
+
+        const MAX_ERRORS_IN_A_ROW: u16 = 50;
+        let mut errors_in_a_row = 0;
+        loop {
+            match NavSysSpsEntry::from_reader(reader) {
+                Ok((entry, bytes_read)) => {
+                    errors_in_a_row = 0;
+                    let ts = entry.timestamp();
+                    if let Some(first_ts) = &mut first_timestamp {
+                        if ts < *first_ts {
+                            first_timestamp = Some(ts);
+                        }
+                    } else {
+                        first_timestamp = Some(ts);
+                    }
+                    entries.push(entry);
+                    total_bytes_read += bytes_read;
+                }
+                Err(e) => {
+                    errors_in_a_row += 1;
+                    if e.kind() == io::ErrorKind::UnexpectedEof {
+                        break;
+                    } else if errors_in_a_row == MAX_ERRORS_IN_A_ROW {
+                        log::warn!("Max errors in a row reached, breaking");
+                        break;
+                    } else {
+                        // Continue on error
+                        log::warn!("Failed parsing log entry: {e}");
+                    }
+                }
+            }
+        }
 
         let raw_plots = Self::build_raw_plots(&entries);
 
         Ok((
             Self {
-                header,
                 entries,
                 raw_plots,
+                first_timestamp: first_timestamp
+                    .expect("invalid condition, no first timestamp in dataset"),
             },
             total_bytes_read,
         ))
@@ -405,13 +434,14 @@ impl Parseable for NavSysSps {
 
     fn is_buf_valid(buf: &[u8]) -> bool {
         let mut reader = BufReader::new(buf);
-        NavSysSpsHeader::from_reader(&mut reader).is_ok()
+        let is_valid_tilt_sensor = Self::is_reader_valid_tilt_sensor(&mut reader);
+        MagSps::is_buf_valid(buf) || Wasp200Sps::is_buf_valid(buf) || is_valid_tilt_sensor
     }
 }
 
-impl GitMetadata for NavSysSps {
+impl GitMetadata for NavSysSpsKitchenSink {
     fn project_version(&self) -> Option<String> {
-        Some(self.header.software_revision().to_owned())
+        None
     }
 
     fn git_short_sha(&self) -> Option<String> {
@@ -430,132 +460,36 @@ impl GitMetadata for NavSysSps {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use plotinator_test_util::*;
-
-    // Example Navsys.sps data with a header and some data taken from the middle of an actual Navsys.sps log
-    // these logs are very long so we don't want to add a real file to the repo test data
-    const TEST_DATA: &str = "\
-VER 3
-MRK 2024 10 03 12 52 42 401 Navsys software rev: Build: 2.0.0.6
-MRK 2024 10 03 12 52 42 417 TiltSensorID : 1459_1458
-MRK 2024 10 03 12 52 42 417 CalAng 1 OffsetY: 0.4950
-MRK 2024 10 03 12 52 42 417 CalAng 1 Y: 3.26488
-MRK 2024 10 03 12 52 42 432 CalAng 1 OffsetX: 0.5099
-MRK 2024 10 03 12 52 42 432 CalAng 1 X: 3.28745
-MRK 2024 10 03 12 52 42 432 CalAng 2 OffsetY: 0.4947
-MRK 2024 10 03 12 52 42 432 CalAng 2 Y: 3.34307
-MRK 2024 10 03 12 52 42 432 CalAng 2 OffsetX: 0.5047
-MRK 2024 10 03 12 52 42 432 CalAng 2 X: 3.30737
-MA1 2024 10 03 13 33 47 035 49966.2649
-TL1 2024 10 03 13 33 46 769 2.28 2.60
-MA1 2024 10 03 13 33 47 055 49966.1729
-MA1 2024 10 03 13 33 47 075 49966.1913
-MA1 2024 10 03 13 33 47 095 49966.3017
-MA1 2024 10 03 13 33 47 115 49966.2465
-MA1 2024 10 03 13 33 47 135 49966.1545
-HE2 2024 10 03 13 33 46 878 213.78
-HE2 2024 10 03 13 33 46 878 214.00
-HE2 2024 10 03 13 33 46 878 214.22
-HE2 2024 10 03 13 33 46 878 214.56
-HE2 2024 10 03 13 33 46 878 214.90
-HE2 2024 10 03 13 33 46 878 215.15
-MA1 2024 10 03 13 33 47 155 49966.3569
-MA1 2024 10 03 13 33 47 175 49966.1729
-MA1 2024 10 03 13 33 47 195 49966.2281
-HE1 2024 10 03 13 33 46 940 99999.99
-HE1 2024 10 03 13 33 46 940 201.62
-HE1 2024 10 03 13 33 46 940 99999.99
-HE1 2024 10 03 13 33 46 940 202.17
-HE1 2024 10 03 13 33 46 940 210.32
-HE1 2024 10 03 13 33 46 940 211.02
-MA1 2024 10 03 13 33 47 215 49966.1729
-MA1 2024 10 03 13 33 47 235 49966.1361
-GP2 2024 10 03 13 33 46 971 5339.76758 910.84021 13:33:47.000 19 WGS84 130.3 0.8 1.3 1.5 193.1
-GP1 2024 10 03 13 33 46 971 5339.76660 910.84003 13:33:47.000 16 WGS84 130.1 0.7 1.0 1.2 193.6
-TL2 2024 10 03 13 33 46 971 2.80 2.34
-MA1 2024 10 03 13 33 47 255 49966.1729
-MA1 2024 10 03 13 33 47 275 49966.3509
-MA1 2024 10 03 13 33 47 295 49966.1913
-MA1 2024 10 03 13 33 47 315 49966.2833
-MA1 2024 10 03 13 33 47 335 49966.1729
-HE2 2024 10 03 13 33 47 081 99999.99
-HE2 2024 10 03 13 33 47 081 99999.99
-HE2 2024 10 03 13 33 47 081 99999.99
-HE2 2024 10 03 13 33 47 081 99999.99
-HE2 2024 10 03 13 33 47 081 99999.99
-HE2 2024 10 03 13 33 47 081 99999.99
-MA1 2024 10 03 13 33 47 355 49966.2281
-MA1 2024 10 03 13 33 47 375 49966.2097
-MA1 2024 10 03 13 33 47 395 49966.2833
-MA1 2024 10 03 13 33 47 415 49966.1545
-HE1 2024 10 03 13 33 47 159 99999.99
-HE1 2024 10 03 13 33 47 159 211.44
-HE1 2024 10 03 13 33 47 159 211.31
-HE1 2024 10 03 13 33 47 159 211.55
-HE1 2024 10 03 13 33 47 159 212.09
-HE1 2024 10 03 13 33 47 159 212.54
-HE1 2024 10 03 13 33 47 159 212.23
-MA1 2024 10 03 13 33 47 435 49966.2649
-MA1 2024 10 03 13 33 47 455 49966.0441
-MA1 2024 10 03 13 33 47 475 49966.2465
-MA1 2024 10 03 13 33 47 495 49966.2465
-MA1 2024 10 03 13 33 47 515 49966.1729
-MA1 2024 10 03 13 33 47 535 49966.0257
-TL1 2024 10 03 13 33 47 283 0.90 5.70
-MA1 2024 10 03 13 33 47 555 49966.2833
-HE2 2024 10 03 13 33 47 299 216.50
-HE2 2024 10 03 13 33 47 299 216.48
-HE2 2024 10 03 13 33 47 299 216.64
-HE2 2024 10 03 13 33 47 299 217.18
-HE2 2024 10 03 13 33 47 299 217.55
-HE2 2024 10 03 13 33 47 299 217.85
-HE2 2024 10 03 13 33 47 299 99999.99
-MA1 2024 10 03 13 33 47 575 49966.1301
-MA1 2024 10 03 13 33 47 595 49966.2649
-MA1 2024 10 03 13 33 47 615 49966.1545
-MA1 2024 10 03 13 33 47 635 49966.3569
-HE1 2024 10 03 13 33 47 377 212.36
-HE1 2024 10 03 13 33 47 377 212.67
-HE1 2024 10 03 13 33 47 377 212.90
-HE1 2024 10 03 13 33 47 377 212.99
-HE1 2024 10 03 13 33 47 377 213.29
-HE1 2024 10 03 13 33 47 377 213.50
-HE1 2024 10 03 13 33 47 377 213.77
-MA1 2024 10 03 13 33 47 655 49966.0257
-MA1 2024 10 03 13 33 47 675 49966.3201
-MA1 2024 10 03 13 33 47 695 49966.0993
-MA1 2024 10 03 13 33 47 715 49966.1913
-MA1 2024 10 03 13 33 47 735 49966.0625
-MA1 2024 10 03 13 33 47 755 49966.3201
-TL2 2024 10 03 13 33 47 486 0.25 6.78
-MA1 2024 10 03 13 33 47 775 49966.1361
-HE2 2024 10 03 13 33 47 517 99999.99
-HE2 2024 10 03 13 33 47 517 99999.99
-HE2 2024 10 03 13 33 47 517 99999.99
-HE2 2024 10 03 13 33 47 517 219.06
-HE2 2024 10 03 13 33 47 517 219.23
-HE2 2024 10 03 13 33 47 517 219.58
-HE2 2024 10 03 13 33 47 517 219.96
-MA1 2024 10 03 13 33 47 795 49966.1545
-MA1 2024 10 03 13 33 47 815 49965.9337
-MA1 2024 10 03 13 33 47 835 49966.2465
-MA1 2024 10 03 13 33 47 855 49966.0993
-HE1 2024 10 03 13 33 47 595 213.96
-HE1 2024 10 03 13 33 47 595 214.04
-HE1 2024 10 03 13 33 47 595 214.33
-HE1 2024 10 03 13 33 47 595 214.53
-";
+    use plotinator_test_util::{
+        test_file_defs::navsys_kitchen_sink::{
+            NAVSYS_SPS_KITCHEN_SINK_BYTES, navsys_sps_kitchen_sink,
+        },
+        *,
+    };
 
     #[test]
-    fn test_parse_navsys_log() -> TestResult {
-        let mut cursor = io::Cursor::new(TEST_DATA);
-        let (navsys, bytes_read) = NavSysSps::from_reader(&mut cursor)?;
+    fn test_is_file_valid() {
+        let is_valid = NavSysSpsKitchenSink::file_is_valid(&navsys_sps_kitchen_sink());
+        assert!(is_valid);
+    }
 
-        assert_eq!(bytes_read, TEST_DATA.len());
+    #[test]
+    fn test_is_file_valid_bifrost_h5_not_valid() {
+        let is_valid = NavSysSpsKitchenSink::file_is_valid(&bifrost_current());
+        assert!(!is_valid);
+    }
 
-        assert_eq!(navsys.header.software_revision(), "Build: 2.0.0.6");
-        assert_eq!(navsys.header.tilt_sensor_id(), "1459_1458");
-        assert_eq!(navsys.entries.len(), 98);
+    #[test]
+    fn test_parse_navsys_kitchen_sink() -> TestResult {
+        let mut test_data = NAVSYS_SPS_KITCHEN_SINK_BYTES;
+        let (navsys, bytes_read) = NavSysSpsKitchenSink::from_reader(&mut test_data)?;
+
+        if cfg!(target_os = "windows") {
+            assert_eq!(bytes_read, 15556);
+        } else {
+            assert_eq!(bytes_read, 15119);
+        }
+        assert_eq!(navsys.entries.len(), 437);
 
         Ok(())
     }