+++ crates/plotinator-test-util/src/test_file_defs.rs
@@ -54,5 +54,6 @@ pub mod frame_inclinometers;
 pub mod frame_magnetometer;
 pub mod legacy_generator;
 pub mod mbed_motor_control;
+pub mod navsys_kitchen_sink;
 pub mod njord_ins;
 pub mod wasp200;