+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for `Frame Inclinometers` HDF5 files
 - Support for `Frame Magnetometer` HDF5 files
+- Support for standalone "kitchen sink" SPS files with any valid Navsys.sps entries
 
 ### Changed